+++ src/core/project_archive.rs
@@ -268,13 +268,3 @@ impl ProjectArchive {
             .collect()
     }
 }
-
-// #[cfg(test)]
-// mod test {
-//     use super::*;
-
-//     #[test]
-//     fn write_and_read() {
-
-//     }
-// }